Switch all embedded DSO handling to use mkstemp.
This switches all internal use of shm based embedded DSO handling
to mkstemp based embedded DSO handling. An issue was identified in
testing where a /dev/shm mount could have the `noexec` flag set
(seemingly quite common in docker containers) and as a result of
this embedded DSOs extracted to this location could not be
executed. Resolving this will take some time and so as a temporary
measure the aforementioned switch is being made.